requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.61.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.6
cyclopts==3.22.5
defog==1.2.6
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.11.2
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.29.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
llama_cpp_python==0.3.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.3
mdurl==0.1.2
mistralai==1.9.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
packaging==25.0
pandas==2.3.1
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
portalocker==3.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pwinput==1.0.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlcoder @ file:///home/tensor/coding/vscode/sqlcoder
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.10.0
together==1.5.23
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
triton==3.4.0
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.1
yarl==1.20.1
